code commit - base on sybase

+++ spring-boot-mybatis/src/main/java/org/example/mapper/StudentMapper.java
@@ -5,13 +5,13 @@ import org.example.domain.entity.Student;
 import java.util.List;
 
 public interface StudentMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Student record);
 
     int insertSelective(Student record);
 
-    Student selectByPrimaryKey(Integer id);
+    Student selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Student record);
